remove /manage paths from my-drops
